Attivati vari avvisi specifici, corretti un po' di avvisi e aggiunte un po' di keyword qua e là al fine di rendere il codice più esplicito.

+++ Codice/Audio C++/CostantiEdAltro.h
@@ -134,7 +134,7 @@ inline size_t DaSecondiACampioni(double secondi)
 
 inline constexpr double DaCampioniASecondi(size_t campioni)
 {
-    return campioni / Costanti::FrequenzaCampionamento;
+    return static_cast<double>(campioni) / Costanti::FrequenzaCampionamento;
 }
 
 inline constexpr double DaSmussamentoAGuadagno(double valore)
